Fucking up the backend :D

+++ tikoht/src/main/java/group32/tikoht/model/Tuntityo.java
@@ -8,12 +8,12 @@ public class Tuntityo {
 
     private final String tyontyyppi;
     @NotBlank
-    private final Integer hinta;
+    private final Double hinta;
     @NotBlank
     private final Integer alv;
 
     public Tuntityo( @JsonProperty("tyontyyppi") String tyontyyppi,
-                    @JsonProperty("hinta") Integer hinta,
+                    @JsonProperty("hinta") Double hinta,
                     @JsonProperty("alv") Integer alv) {
         this.tyontyyppi = tyontyyppi;
         this.hinta = hinta;
@@ -24,7 +24,7 @@ public class Tuntityo {
         return tyontyyppi;
     }
 
-    public Integer getHinta() {
+    public Double getHinta() {
         return hinta;
     }
